+++ source/DSCResources/DSC_DefaultGatewayAddress/DSC_DefaultGatewayAddress.psm1
@@ -298,7 +298,7 @@ function Assert-ResourceProperty
 
     if ($Address)
     {
-        Test-IPAddress -Address $Address -AddressFamily $AddressFamily
+        Assert-IPAddress -Address $Address -AddressFamily $AddressFamily
     }
 } # Assert-ResourceProperty